Query data from the Altmetric.com API

Query data from the Altmetric.com API

altmetrics(oid = NULL, id = NULL, doi = NULL, pmid = NULL,
  arXiv = NULL, isbn = NULL, uri = NULL,
  apikey = getOption("altmetricKey"), foptions = list(), ...)

Arguments

oid
An object ID

id
An altmetric.com id for a scholarly paper

doi
A DOI

pmid
A pubmed ID

arXiv
An Arxiv id.

isbn
ISBN

uri
Any URI

apikey
Your API `key`. By default the package ships with a key, but mostly as a demo. If the key becomes overused, then it is likely that you will start to see API limit errors

foptions
Additional options for `httr`

...
additional options

Examples

not_run({
  altmetrics(doi ='10.1038/480426a')
  # For ISBNs
  ib <- altmetrics(isbn = "978-3-319-25557-6")
})